+++ com.unity.build-report-inspector/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [0.4.2-preview] - 2025-02-21
+
+-Add Content Summary tab
+-Fix bug when Asset Bundle / Player build type was not changing when switching selected BuildReport
+-Fix for issue 21 - exception related to sprite altas path parsing.
 
 ## [0.4.1-preview] - 2025-02-18